+++ NEWS.md
@@ -1,3 +1,10 @@
+plumber 0.2.4
+--------------------------------------------------------------------------------
+* Add a filter which parses and sets req$cookies to be a list corresponding to
+  the cookies provided with the request.
+* Bug Fix: conver non-character arguments in setCookie to character before URL-
+  encoding.
+
 plumber 0.2.3
 --------------------------------------------------------------------------------
 * Set options(warn=1) during execution of user code so that warnings are